requirements.txt
slack-sdk>=3.21.0
python-dotenv>=1.0.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
fastapi>=0.115.0
uvicorn>=0.34.0
supabase>=2.0.0
apscheduler>=3.10.0
